+++ src/main/java/ac/grim/grimac/api/AbstractCheck.java
@@ -1,4 +1,4 @@
-package ac.grim.grimac;
+package ac.grim.grimac.api;
 
 public interface AbstractCheck {
